% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{get_berlinerisch}
\alias{get_berlinerisch}
\title{get_berlinerisch}
\usage{
get_berlinerisch(txt)
}
\arguments{
\item{txt}{provide a text you want to translate from German to Berlin dialect}
}
\description{
Translates German to Berlin dialect
}
